backend/pyproject.toml
[tool.poetry]
name = "vrp-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
requests = "^2.28.1"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.26.0"
SQLAlchemy = "^1.4.44"
databases = "^0.6.1"
gunicorn = "^20.1.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.2.0"}
GeoAlchemy2 = "^0.10.2"
sqlalchemy_serializer = "^1.4"
shapely = "*"
shapely-geojson = "*"
pandas = "^1.5.1"
openrouteservice = "*"
ortools = "^9.4.1874"
geopandas = "^0.10.2"
httpx = "^0.23.1"
pyvroom = "^0.0.14"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pytest = "^7.1.3"
ipython = "^8.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.cruft]
skip = [".env"]
